Add files via upload
Added an empty text check before scoring a candidate machine-generated summary because the Bleu scorer doesn't accept an empty text or a text full of spaces as a candidate 
When a method produces an empty text file the score of that candidate summary should be 0. Since we divide by the number of files in the produced folder, the average method score isn't affected when we don't add the empty text score.
We also added two more abstractive models for the evaluation.